Refactor Arma 3 scenario autocomplete into separate component
