+++ frontend/src/components/servers/Arma3ScenarioRotationForm.tsx
@@ -1,31 +1,21 @@
 import { Add, Delete } from "@mui/icons-material";
-import { Autocomplete, AutocompleteValue, Box, Button, Card, Grid, IconButton, List, ListItem, ListItemText, Stack, TextField, Typography } from "@mui/material";
+import { AutocompleteValue, Button, Card, Grid, IconButton, List, ListItem, ListItemText, Stack, Typography } from "@mui/material";
 import { FormikProps } from "formik";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { Arma3ScenarioDto } from "../../dtos/Arma3ScenarioDto";
 import { Arma3ServerDto } from "../../dtos/ServerDto";
-import { getScenarios } from "../../services/scenarioService";
+import { Arma3ScenariosAutocomplete } from "./Arma3ScenariosAutocomplete";
 
 type Arma3ScenarioRotationFormProps = {
     formik: FormikProps<Arma3ServerDto>
 }
 
 export const Arma3ScenarioRotationForm = ({ formik }: Arma3ScenarioRotationFormProps) => {
-    const [scenarios, setScenarios] = useState<Array<Arma3ScenarioDto>>([]);
     const [selectedScenario, setSelectedScenario] = useState<string>();
-
-    useEffect(() => {
-        const fetchScenarios = async () => {
-            const { data: scenariosDto } = await getScenarios();
-            setScenarios(scenariosDto.scenarios);
-        };
-
-        fetchScenarios();
-        setSelectedScenario("");
-    }, []);
 
     const setScenarioAutocomplete = (_: any, value: AutocompleteValue<Arma3ScenarioDto | string, false, false, true> | null): void => {
         if (!value) {
+            setSelectedScenario("");
             return;
         }
         setSelectedScenario(typeof value === "string" ? value : value.name);
@@ -47,14 +37,6 @@ export const Arma3ScenarioRotationForm = ({ formik }: Arma3ScenarioRotationFormP
         formik.setFieldValue("scenarios", [...formik.values.scenarios].filter(scenario => scenario !== scenarioName));
     }
 
-    const getScenarioDisplayName = (scenario: Arma3ScenarioDto) => {
-        return scenario.name; // TOOD: Can make pretty later
-    };
-
-    const handleBlur = (e: React.FocusEvent<HTMLInputElement>) => {
-        setScenarioAutocomplete(null, e.target.value);
-    };
-
     return (
         <Grid container spacing={1}>
             <Grid item xs={12}>
@@ -67,44 +49,12 @@ export const Arma3ScenarioRotationForm = ({ formik }: Arma3ScenarioRotationFormP
                     alignItems="center"
                     spacing={1}
                     sx={{ flexGrow: 1 }}>
-                    <Autocomplete // TODO: Move this to it's own component?
-                        id="scenario-select"
-                        fullWidth
-                        freeSolo
-                        selectOnFocus
-                        onBlur={handleBlur}
-                        handleHomeEndKeys
-                        options={scenarios}
-                        autoHighlight
-                        getOptionLabel={(option) => {
-                            if (typeof option === "string")
-                                return option;
-                            return option.name;
-                        }}
-                        onChange={setScenarioAutocomplete}
-                        renderOption={(props, option) => (
-                            <Box component="li"  {...props}>
-                                {getScenarioDisplayName(option)}
-                            </Box>
-                        )}
-                        renderInput={(params) => (
-                            <TextField
-                                {...params}
-                                name="scenarioName"
-                                label="Scenario name"
-                                inputProps={{
-                                    ...params.inputProps,
-                                    autoComplete: 'new-password'
-                                }}
-                            />
-                        )}
-                    />
-
+                    <Arma3ScenariosAutocomplete onChange={setScenarioAutocomplete} />
                     <Button
                         variant="contained"
                         startIcon={<Add />}
                         onClick={() => addScenario(selectedScenario ?? "")}
-                        disabled={Boolean(formik.values.scenarios.find((scenario) => scenario === selectedScenario))}>
+                        disabled={selectedScenario ? Boolean(formik.values.scenarios.find((scenario) => scenario === selectedScenario)) : true}>
                         Add
                     </Button>
                 </Stack>
